feat: further prune some gateway device types that do not need to be known by core
