+++ meteor/client/ui/Settings/components/ConfigManifestOAuthFlow.tsx
@@ -1,12 +1,9 @@
 import * as React from 'react'
 import { translate } from 'react-i18next'
-import {
-	PeripheralDevice,
-	SpreadsheetDevice,
-} from '../../../../lib/collections/PeripheralDevices'
+import { PeripheralDevice } from '../../../../lib/collections/PeripheralDevices'
 import { EditAttribute, EditAttributeBase } from '../../../lib/EditAttribute'
 import { Translated } from '../../../lib/ReactMeteorData/react-meteor-data'
-import { SpreadsheetDeviceSettings } from '../../../../lib/collections/PeripheralDeviceSettings/spreadsheet'
+import { IngestDeviceSettings } from '../../../../lib/collections/PeripheralDeviceSettings/ingestDevice'
 import { NotificationCenter, Notification, NoticeLevel } from '../../../lib/notifications/notifications'
 import { PeripheralDeviceAPI } from '../../../../lib/api/peripheralDevice'
 import { fetchFrom } from '../../../lib/lib'
@@ -109,8 +106,8 @@ export const ConfigManifestOAuthFlowComponent = translate()(class ConfigManifest
 	}
 	render () {
 		const { t } = this.props
-		let settings = (this.props.device.settings || {}) as SpreadsheetDeviceSettings
-		let device = this.props.device as SpreadsheetDevice
+		let settings = (this.props.device.settings || {}) as IngestDeviceSettings
+		let device = this.props.device
 		return (<div>
 			<div className='mod mvs mhn'>
 				<label className='field'>
