--- meteor/lib/collections/PeripheralDeviceSettings/iNews.ts
@@ -1,16 +1,0 @@
-export interface INewsDeviceSettings {
-	hosts: Array<INewsHost>
-	user: string
-	password: string
-	queues: Array<INewsQueue>
-}
-
-export interface INewsHost {
-	_id: string
-	host: string
-}
-
-export interface INewsQueue {
-	_id: string
-	queue: string
-}
