--- meteor/lib/collections/PeripheralDeviceSettings/mosDevice.ts
@@ -1,14 +1,0 @@
-export interface MosDeviceSettings {
-	mosId: string,
-	devices: {
-		[deviceId: string]: MosDeviceSettingsDevice
-	}
-}
-export interface MosDeviceSettingsDevice {
-	primary: MosDeviceSettingsDeviceOptions
-	secondary?: MosDeviceSettingsDeviceOptions
-}
-export interface MosDeviceSettingsDeviceOptions {
-	id: string
-	host: string
-}
