+++ meteor/lib/collections/PeripheralDeviceSettings/playoutDevice.ts
@@ -1,13 +1,11 @@
 import { TSR } from 'tv-automation-sofie-blueprints-integration'
 
+/** 
+ * The basic PlayoutDevice settings structure.
+ * Note: playout-gateway will likely have more than this here, but this is that core needs to know about
+ */
 export interface PlayoutDeviceSettings {
 	devices: {
 		[deviceId: string]: TSR.DeviceOptionsAny
 	}
-	mediaScanner: {
-		host: string
-		port: number
-	}
-	multiThreading?: boolean
-	multiThreadedResolver?: boolean
 }
