+++ meteor/lib/collections/PeripheralDevices.ts
@@ -3,11 +3,8 @@ import { Time, registerCollection, ProtectedString } from '../lib'
 import { TransformedCollection } from '../typings/meteor'
 import { Meteor } from 'meteor/meteor'
 
-import { MediaManagerDeviceSettings } from './PeripheralDeviceSettings/mediaManager'
 import { PlayoutDeviceSettings } from './PeripheralDeviceSettings/playoutDevice'
-import { MosDeviceSettings } from './PeripheralDeviceSettings/mosDevice'
-import { SpreadsheetDeviceSettings, SpreadsheetDeviceSecretSettings } from './PeripheralDeviceSettings/spreadsheet'
-import { INewsDeviceSettings } from './PeripheralDeviceSettings/iNews'
+import { IngestDeviceSettings, IngestDeviceSecretSettings } from './PeripheralDeviceSettings/ingestDevice'
 import { createMongoCollection } from './lib'
 import { DeviceConfigManifest } from '../api/deviceConfig'
 import { StudioId } from './Studios'
@@ -46,51 +43,20 @@ export interface PeripheralDevice {
 
 	token: string
 
-	settings?: MosDeviceSettings | PlayoutDeviceSettings | MediaManagerDeviceSettings | SpreadsheetDeviceSettings | INewsDeviceSettings
+	settings?: PlayoutDeviceSettings | IngestDeviceSettings | { [key: string]: any }
 
-	secretSettings?: any | SpreadsheetDeviceSecretSettings
+	secretSettings?: IngestDeviceSecretSettings | { [key: string]: any }
 
 	/** Ignore this device when computing status in the GUI (other status reports are unaffected) */
 	ignore?: boolean
 
 	configManifest: DeviceConfigManifest
-}
-
-export interface MosParentDevice extends PeripheralDevice {
-	category: PeripheralDeviceAPI.DeviceCategory.INGEST,
-	type: PeripheralDeviceAPI.DeviceType.MOS,
-	subType: PeripheralDeviceAPI.SUBTYPE_PROCESS,
-	settings?: MosDeviceSettings
-	secretSettings: undefined
+	
+	/** If this is an ingest gateway, the last tiem data was received */
 	lastDataReceived?: Time
-}
-export interface PlayoutParentDevice extends PeripheralDevice {
-	category: PeripheralDeviceAPI.DeviceCategory.PLAYOUT,
-	type: PeripheralDeviceAPI.DeviceType.PLAYOUT,
-	subType: PeripheralDeviceAPI.SUBTYPE_PROCESS,
-	secretSettings: undefined
-	settings?: PlayoutDeviceSettings
-}
-export interface MediaManagerDevice extends PeripheralDevice {
-	category: PeripheralDeviceAPI.DeviceCategory.MEDIA_MANAGER,
-	type: PeripheralDeviceAPI.DeviceType.MEDIA_MANAGER,
-	subType: PeripheralDeviceAPI.SUBTYPE_PROCESS,
-	secretSettings: undefined
-	settings?: MediaManagerDeviceSettings
-}
-export interface SpreadsheetDevice extends PeripheralDevice {
-	category: PeripheralDeviceAPI.DeviceCategory.INGEST,
-	type: PeripheralDeviceAPI.DeviceType.SPREADSHEET,
-	subType: PeripheralDeviceAPI.SUBTYPE_PROCESS,
-	settings?: SpreadsheetDeviceSettings
-	secretSettings?: SpreadsheetDeviceSecretSettings
+	
+	/** If an ingest device performing an oauth flow */
 	accessTokenUrl?: string
-}
-export interface INewsDevice extends PeripheralDevice {
-	category: PeripheralDeviceAPI.DeviceCategory.INGEST,
-	type: PeripheralDeviceAPI.DeviceType.INEWS,
-	subType: PeripheralDeviceAPI.SUBTYPE_PROCESS,
-	settings?: INewsDeviceSettings
 }
 
 export const PeripheralDevices: TransformedCollection<PeripheralDevice, PeripheralDevice>
